pyproject.toml
[project]
name = "cloudflare-ai-mcp"
version = "0.1.0"
description = """
A Model Context Protocol (MCP) server for Cloudflare's AI API that enables advanced image generation capabilities. \
This package provides a robust interface for generating images using multiple Stable Diffusion models simultaneously, \
with features including AI-powered prompt generation, batch operations, and comprehensive error handling. \
Designed for seamless integration with Claude and other MCP-compatible clients.

Key features:
• Multiple model support for Stable Diffusion
• Concurrent image generation
• AI-powered prompt enhancement
• Batch processing capabilities
• Comprehensive error handling
• Production-ready with monitoring support
"""
authors = [
    {name = "ouraihub", email = "ouraihub@gmail.com"},
]
dependencies = [
    "streamlit>=1.31.1",
    "requests>=2.31.0",
    "Pillow>=10.2.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.3",
    "uvicorn>=0.15.0",
    "cloudflare>=2.8.15",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=6.0.0",
    "black>=22.3.0",
    "isort>=5.10.1",
    "mypy>=0.982",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cloudflare_ai"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

keywords = ["cloudflare", "ai", "image-generation", "mcp", "stable-diffusion", "claude"] 